1/18/2018
Fixed the redundancy, I was populating a 2D array while also populating an adjacency list. Now only the adjacency list is populated. Next thing I need to do is figure out how to update the list as a tile changes from active, deactivate, start, or goal. Two obvious choices are updating the list as the tile changes, or update the list with a button after the user is done setting up the board.