Final Level Editor change
I ended up using a 2D array to represent the board. No reason to use an adjacency list because there is a uniform number of edges, and if there is a wall there is no hassle of disconnecting all of the edges for both the wall and its neighbors when it has to change. The level editor is finally complete, I will now be able to focus on the AI portion.